seller_service/requirements.txt
fastapi[standard]
sqlmodel
requests
python-jose 
fastapi[all]